requirements.txt
asgiref==3.9.1
Django==4.2.7
django-crontab==0.7.1
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
